third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/forms/the-option-element/option-selected-expected.txt

This is a testharness.js-based test.
FAIL not dirty assert_equals: expected false but got true
FAIL dirty, selected assert_equals: expected false but got true
FAIL dirty, not selected assert_equals: expected false but got true
Harness: the test ran to completion.
